Split JS Handle mapping
